refactor : CustomException 리팩토링
- 기존에 도메인별로 TemplateException 하나만 구현하고 해당 Exception으로 모든 템플릿 관련된 예외처리를 하는 방향으로 했는데, 이렇게 되면 무슨 예외인지 직관적이지 않아서 상세 Exception 구현하도록 정함
- 예) TemplateNotFoundException, MyTemplateNotMatchedException 등 상세 예외내용에 대한 Exception 생성
- 그리고 기존에 global/error/exception -> 해당 경로에 모든 CustomException을 생성했는데 각 도메인별로 생성하도록 패키지 구조 변경
- 예) domain/template/exception/ -> TemplateNotFoundException 등 존재

Ref : #110
