+++ src/main/java/com/fastcampus/jober/domain/template/exception/TemplateException.java
@@ -1,6 +1,7 @@
-package com.fastcampus.jober.global.error.exception;
+package com.fastcampus.jober.domain.template.exception;
 
 import com.fastcampus.jober.global.constant.ErrorCode;
+import com.fastcampus.jober.global.error.exception.DomainException;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor
