--- src/main/java/com/fastcampus/jober/global/error/exception/MyTemplateException.java
@@ -1,12 +1,0 @@
-package com.fastcampus.jober.global.error.exception;
-
-import com.fastcampus.jober.global.constant.ErrorCode;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-public class MyTemplateException extends DomainException {
-
-    public MyTemplateException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
